完成科室的增删查改

+++ src/main/java/com/markmzy/controller/UserController.java
@@ -1,18 +1,17 @@
 package com.markmzy.controller;
 
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.markmzy.model.User;
+import com.markmzy.service.IUserService;
+import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
-import org.springframework.web.bind.annotation.*;
-import com.markmzy.service.IUserService;
-import com.markmzy.model.User;
-import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.baomidou.mybatisplus.core.metadata.IPage;
+import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
-import org.springframework.web.bind.annotation.RestController;
 
 /**
  * <p>
@@ -25,7 +24,8 @@ import org.springframework.web.bind.annotation.RestController;
 @Api(tags = {"用户信息表"})
 @RestController
 @RequestMapping("/user")
-public class UserController {
+public class UserController
+{
 
     private Logger log = LoggerFactory.getLogger(getClass());
 
@@ -35,36 +35,41 @@ public class UserController {
 
     @ApiOperation(value = "新增用户信息表")
     @PostMapping()
-    public int add(@RequestBody User user){
+    public int add(@RequestBody User user)
+    {
         return userService.add(user);
     }
 
     @ApiOperation(value = "删除用户信息表")
     @DeleteMapping("{id}")
-    public int delete(@PathVariable("id") Long id){
+    public int delete(@PathVariable("id") Integer id)
+    {
         return userService.delete(id);
     }
 
     @ApiOperation(value = "更新用户信息表")
     @PutMapping()
-    public int update(@RequestBody User user){
+    public int update(@RequestBody User user)
+    {
         return userService.updateData(user);
     }
 
     @ApiOperation(value = "查询用户信息表分页数据")
     @ApiImplicitParams({
-        @ApiImplicitParam(name = "page", value = "页码"),
-        @ApiImplicitParam(name = "pageCount", value = "每页条数")
+            @ApiImplicitParam(name = "page", value = "页码"),
+            @ApiImplicitParam(name = "pageCount", value = "每页条数")
     })
     @GetMapping()
     public IPage<User> findListByPage(@RequestParam Integer page,
-                                   @RequestParam Integer pageCount){
+                                      @RequestParam Integer pageCount)
+    {
         return userService.findListByPage(page, pageCount);
     }
 
     @ApiOperation(value = "id查询用户信息表")
     @GetMapping("{id}")
-    public User findById(@PathVariable Long id){
+    public User findById(@PathVariable Integer id)
+    {
         return userService.findById(id);
     }
